fix(system): touch id, password validator

+++ wallet/CBWallet/Helpers/NSString+Password.h
@@ -10,7 +10,6 @@
 
 @interface NSString (Password)
 
-/// max 100
-- (NSInteger)passwordStrength;
+- (double)passwordStrength;
 
 @end
